SMILE-based Model/requirements.txt
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.3
rdkit==2024.3.5
seaborn==0.13.2
scikit-learn==1.5.2
tabulate==0.9.0
pyyaml==6.0.2